gtk3/gtk2: fix mismatched types for keyboard config temp storage
fallout from c68cc14, closes #558

+++ desmume/src/frontend/posix/gtk/main.cpp
@@ -353,7 +353,7 @@ struct modify_key_ctx {
 };
 
 static u16 keys_latch = 0;
-u16 Keypad_Temp[NB_KEYS];
+u32 Keypad_Temp[NB_KEYS];
 
 class configured_features : public CommandLine
 {
